+++ src/main/scala/com/trafficland/augmentsbt/rpm/RPMPlugin.scala
@@ -1,20 +1,28 @@
 package com.trafficland.augmentsbt.rpm
 
-import com.typesafe.sbt.packager.rpm.{RpmKeys, RpmPlugin}
-import sbt._
-import sbt.Keys._
-import com.typesafe.sbt.packager.linux.LinuxSymlink
+import java.nio.charset.Charset
 
-import scala.collection._
-import com.typesafe.sbt.packager.Keys._
-import com.typesafe.sbt.SbtNativePackager._
 import com.trafficland.augmentsbt.rpm.Keys._
+import com.typesafe.sbt.SbtNativePackager._
+import com.typesafe.sbt.packager.Keys._
+import com.typesafe.sbt.packager.archetypes.{JavaServerAppPackaging, TemplateWriter}
+import com.typesafe.sbt.packager.linux.LinuxSymlink
+import com.typesafe.sbt.packager.rpm.RpmPlugin.autoImport.RpmConstants._
+import com.typesafe.sbt.packager.rpm.{RpmKeys, RpmPlugin}
+import sbt.Keys._
+import sbt._
+
+import scala.util.control.Exception.nonFatalCatch
 
 object RPMPlugin extends AutoPlugin with RpmKeys {
 
-  override def requires = RpmPlugin
+  override def requires: Plugins = RpmPlugin && JavaServerAppPackaging
 
-  override lazy val projectSettings =
+  object Names {
+    val ManageDaemonAccounts = "manage_daemon_accounts"
+  }
+
+  override lazy val projectSettings: Seq[Def.Setting[_]] =
     Seq(
       name in Rpm <<= name apply { n => n },
       linuxPackageSymlinks <<= (installationDirectory, rpmVendor, name in Rpm) map { (installationDir, v, n) =>
@@ -33,5 +41,39 @@ object RPMPlugin extends AutoPlugin with RpmKeys {
       vendorDirectory <<= rpmVendor apply { rv => "/opt/" + rv },
       installationDirectory <<= (vendorDirectory, destinationDirectory) apply { (vd, dd) => vd + Path.sep + dd },
       destinationDirectory <<= (name in Rpm) apply { n => n }
-    )
+    ) ++ inConfig(Rpm)(Seq(
+      manageDaemonAccounts := false,
+      scriptTemplates in Rpm := defaultScriptTemplates(getClass.getResource("")),
+      linuxScriptReplacements <+= manageDaemonAccounts {
+        Names.ManageDaemonAccounts -> _.toString
+      },
+      maintainerScripts in Rpm := {
+        val scripts = makeMaintainerScripts(scriptTemplates.value, linuxScriptReplacements.value)
+        if (rpmBrpJavaRepackJars.value) {
+          val pre = scripts.getOrElse(Pre, Nil)
+          val skipRepackScript = IO.readStream(skipRepackScriptResource.openStream, Charset.forName("UTF-8"))
+          scripts + (Pre -> (pre :+ skipRepackScript))
+        } else {
+          scripts
+        }
+      }
+    ))
+
+  def defaultScriptTemplates(path: URL): Map[String, URL] = {
+    val predefined = Seq(Pre, Post, Preun, Postun)
+    predefined.foldLeft(Map.empty[String, URL]) { (scripts, name) =>
+      Option(new URL(path, s"$name-template")).collect {
+        case script if nonFatalCatch.opt(script.openConnection()).isDefined =>
+          scripts + (name -> script)
+      }.getOrElse(scripts)
+    }
+  }
+
+  def skipRepackScriptResource: java.net.URL = RpmPlugin.getClass.getResource("brpJavaRepackJar")
+
+  def makeMaintainerScripts(templates: Map[String, URL], replacements: Seq[(String, String)]): Map[String, Seq[String]] = {
+    templates.mapValues { url =>
+      TemplateWriter.generateScript(url, replacements) :: Nil
+    }
+  }
 }